requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-coinpayments==0.1.2
django-cors-headers==2.4.0
django-model-utils==3.1.2
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
pytz==2018.5
requests==2.19.1
uritemplate==3.0.0
urllib3==1.23
djangorestframework-simplejwt==3.3
django-celery==3.2.2
redis==2.10.6
celery==3.1.26.post2

